+++ src/commands/functions-delete.ts
@@ -16,7 +16,6 @@ import * as planner from "../deploy/functions/release/planner";
 import * as fabricator from "../deploy/functions/release/fabricator";
 import * as executor from "../deploy/functions/release/executor";
 import * as reporter from "../deploy/functions/release/reporter";
-import * as containerCleaner from "../deploy/functions/containerCleaner";
 import { getProjectNumber } from "../getProjectNumber";
 
 export const command = new Command("functions:delete [filters...]")
@@ -106,7 +105,4 @@ export const command = new Command("functions:delete [filters...]")
         exit: 1,
       });
     }
-
-    // Clean up image caches too
-    await containerCleaner.cleanupBuildImages([], allEpToDelete);
   });