+++ src/mcp/index.ts
@@ -16,13 +16,15 @@ import { Command } from "../command.js";
 import { requireAuth } from "../requireAuth.js";
 import { Options } from "../options.js";
 import { getProjectId } from "../projectUtils.js";
-import { mcpAuthError, NO_PROJECT_ERROR } from "./errors.js";
+import { mcpAuthError, NO_PROJECT_ERROR, mcpGeminiError } from "./errors.js";
 import { trackGA4 } from "../track.js";
 import { Config } from "../config.js";
 import { loadRC } from "../rc.js";
 import { EmulatorHubClient } from "../emulator/hubClient.js";
 import { Emulators } from "../emulator/types.js";
 import { existsSync } from "node:fs";
+import { ensure, check } from "../ensureApiEnabled.js";
+import * as api from "../api.js";
 
 const SERVER_VERSION = "0.1.0";
 
@@ -174,7 +176,8 @@ export class FirebaseMcpServer {
 
   async getAuthenticatedUser(): Promise<string | null> {
     try {
-      return await requireAuth(await this.resolveOptions());
+      const email = await requireAuth(await this.resolveOptions());
+      return email ?? "Application Default Credentials";
     } catch (e) {
       return null;
     }
@@ -206,6 +209,7 @@ export class FirebaseMcpServer {
     const tool = this.getTool(toolName);
     if (!tool) throw new Error(`Tool '${toolName}' could not be found.`);
 
+    // Check if the current project directory exists.
     if (
       tool.mcp.name !== "firebase_update_environment" && // allow this tool only, to fix the issue
       (!this.cachedProjectRoot || !existsSync(this.cachedProjectRoot))
@@ -214,15 +218,29 @@ export class FirebaseMcpServer {
         `The current project directory '${this.cachedProjectRoot || "<NO PROJECT DIRECTORY FOUND>"}' does not exist. Please use the 'update_firebase_environment' tool to target a different project directory.`,
       );
     }
+
+    // Check if the project ID is set.
     let projectId = await this.getProjectId();
     if (tool.mcp._meta?.requiresProject && !projectId) {
       return NO_PROJECT_ERROR;
     }
     projectId = projectId || "";
 
+    // Check if the user is logged in.
     const accountEmail = await this.getAuthenticatedUser();
     if (tool.mcp._meta?.requiresAuth && !accountEmail) {
       return mcpAuthError();
+    }
+
+    // Check if the tool requires Gemini in Firebase API.
+    if (tool.mcp._meta?.requiresGemini) {
+      if (configstore.get("gemini")) {
+        await ensure(projectId, api.cloudAiCompanionOrigin(), "");
+      } else {
+        if (!(await check(projectId, api.cloudAiCompanionOrigin(), ""))) {
+          return mcpGeminiError(projectId);
+        }
+      }
     }
 
     const options = { projectDir: this.cachedProjectRoot, cwd: this.cachedProjectRoot };