+++ CHANGELOG.md
@@ -1,2 +1,4 @@
-- Add emulator support to firebase storage MCP tools. (#8707)
+- Added validation to check if project ID exists during project creation. (#5233)
+- Add `generate_dataconnect_schema`, `dataconnect_generate_operation`, `firebase_consult_assistant` MCP tools. (#8647)
+- `firebase init dataconnect` is now integrated with Gemini in Firebase API to generate Schema based on description. (#8596)
 - Add user-friendly warning for runtime errors from the GCF API. (#8713)